Move reservation modal to Library / Modals
has been moved so that it is placed with the other modals and displays in a modal layout
