+++ base.scss
@@ -55,6 +55,8 @@
 @import "./src/stories/Library/Modals/modal-text/modal-text";
 @import "./src/stories/Library/Modals/modal-pause/modal-pause";
 @import "./src/stories/Library/Modals/modal-profile/modal-profile";
+@import "./src/stories/Library/modals/modal-reservation/reservation.scss";
+
 @import "./src/stories/Blocks/material-page/material-page";
 @import "./src/stories/Library/Buttons/toggle-button/toggle-button";
 @import "./src/stories/Library/pause-reservation/pause-reservation";
@@ -81,7 +83,6 @@
 @import "./src/stories/Blocks/status-loans/status-loans";
 @import "./src/stories/Blocks/status-userprofile/status-userprofile";
 @import "./src/stories/Blocks/material-manifestation-item/material-manifestation-item";
-@import "./src/stories/Blocks/reservation/reservation.scss";
 
 
 @import "./src/styles/scss/shared";
